#26 fix: Rename date util
